fix(condo): INFRA-499 refactored filters
