+++ apps/condo/domains/contact/hooks/useContactRolesTableColumns.tsx
@@ -4,7 +4,7 @@ import { IntlShape } from 'react-intl/src/types'
 
 import { useIntl } from '@open-condo/next/intl'
 
-import { getFilterIcon } from '@condo/domains/common/components/TableFilter'
+import { getFilterIcon } from '@condo/domains/common/components/Table/Filters'
 import { FiltersMeta, getFilterDropdownByKey } from '@condo/domains/common/utils/filters.utils'
 import { getFilteredValue } from '@condo/domains/common/utils/helpers'
 import { parseQuery } from '@condo/domains/common/utils/tables.utils'
