+++ apps/condo/domains/marketplace/hooks/useMarketplaceInvoicesTableColumns.tsx
@@ -5,8 +5,8 @@ import { useMemo } from 'react'
 import { useIntl } from '@open-condo/next/intl'
 import { Tag } from '@open-condo/ui'
 
+import { getFilterIcon } from '@condo/domains/common/components/Table/Filters'
 import { getDateRender, getTableCellRenderer } from '@condo/domains/common/components/Table/Renders'
-import { getFilterIcon } from '@condo/domains/common/components/TableFilter'
 import { getFilterDropdownByKey } from '@condo/domains/common/utils/filters.utils'
 import { getFilteredValue } from '@condo/domains/common/utils/helpers'
 import { getSorterMap, parseQuery } from '@condo/domains/common/utils/tables.utils'
