+++ apps/condo/domains/organization/hooks/useEmployeeRolesTableColumns.tsx
@@ -11,7 +11,7 @@ import React, { useCallback, useMemo } from 'react'
 import { useIntl } from '@open-condo/next/intl'
 import { Tooltip, Typography } from '@open-condo/ui'
 
-import { getFilterIcon } from '@condo/domains/common/components/TableFilter'
+import { getFilterIcon } from '@condo/domains/common/components/Table/Filters'
 import { FiltersMeta, getFilterDropdownByKey } from '@condo/domains/common/utils/filters.utils'
 import { getFilteredValue } from '@condo/domains/common/utils/helpers'
 import { parseQuery } from '@condo/domains/common/utils/tables.utils'
