+++ apps/condo/domains/scope/hooks/useTableColumns.tsx
@@ -5,8 +5,8 @@ import { useCallback, useMemo } from 'react'
 
 import { useIntl } from '@open-condo/next/intl'
 
+import { getFilterIcon } from '@condo/domains/common/components/Table/Filters'
 import { getTableCellRenderer } from '@condo/domains/common/components/Table/Renders'
-import { getFilterIcon } from '@condo/domains/common/components/TableFilter'
 import { getFilterDropdownByKey } from '@condo/domains/common/utils/filters.utils'
 import { getFilteredValue } from '@condo/domains/common/utils/helpers'
 import { parseQuery } from '@condo/domains/common/utils/tables.utils'
