+++ apps/condo/domains/ticket/hooks/useCallRecordTableColumns.tsx
@@ -8,8 +8,8 @@ import { useOrganization } from '@open-condo/next/organization'
 import { Space, Tag, Tooltip } from '@open-condo/ui'
 import { colors } from '@open-condo/ui/dist/colors'
 
+import { getFilterIcon } from '@condo/domains/common/components/Table/Filters'
 import { getAddressRender, getDateRender, getTableCellRenderer } from '@condo/domains/common/components/Table/Renders'
-import { getFilterIcon } from '@condo/domains/common/components/TableFilter'
 import { useDownloadFileFromServer } from '@condo/domains/common/hooks/useDownloadFileFromServer'
 import { getFilterDropdownByKey } from '@condo/domains/common/utils/filters.utils'
 import { formatPhone, getFilteredValue } from '@condo/domains/common/utils/helpers'
