+++ apps/condo/domains/ticket/hooks/useClientCardTicketTableColumns.tsx
@@ -4,8 +4,8 @@ import { useCallback, useMemo } from 'react'
 
 import { useIntl } from '@open-condo/next/intl'
 
+import { getFilterIcon } from '@condo/domains/common/components/Table/Filters'
 import { getDateRender, getTableCellRenderer } from '@condo/domains/common/components/Table/Renders'
-import { getFilterIcon } from '@condo/domains/common/components/TableFilter'
 import { getFilteredValue } from '@condo/domains/common/utils/helpers'
 import { getSorterMap, parseQuery } from '@condo/domains/common/utils/tables.utils'
 import { TicketComment } from '@condo/domains/ticket/utils/clientSchema'
