+++ apps/condo/domains/ticket/hooks/useIncidentTableColumns.tsx
@@ -10,7 +10,7 @@ import { useCallback, useEffect, useMemo, useState } from 'react'
 
 import { useIntl } from '@open-condo/next/intl'
 
-import { getFilterIcon } from '@condo/domains/common/components/TableFilter'
+import { getFilterIcon } from '@condo/domains/common/components/Table/Filters'
 import { getFilterDropdownByKey } from '@condo/domains/common/utils/filters.utils'
 import { getFilteredValue } from '@condo/domains/common/utils/helpers'
 import { getSorterMap, parseQuery } from '@condo/domains/common/utils/tables.utils'
